KnobControl - clean up imports, App.jsx - refactor Instrument into component, remove Instrument component, Output - change constructor to not take default values as args
